refactor: post 패키지의 controller를 Query와 Command로 분리
